Reverted change.

+++ vdi/vdi_mouse.c
@@ -589,7 +589,7 @@ void vdimouse_init(Vwk * vwk)
     *vblqueue = (LONG)vb_draw;   /* set GEM VBL-routine to vbl_list[0] */
 
     /* Initialize mouse via XBIOS in relative mode */
-    Initmous(1, (LONG)&mouse_params, mouse_int);
+    Initmous(1, (LONG)&mouse_params, (LONG)mouse_int);
 
     kbd_vectors = (struct kbdvecs *)Kbdvbase();
     old_statvec = kbd_vectors->statvec;
